fix: CommentsList에서 댓글 전체 관리, 참여 버튼 컴포넌트 MeetingInfo 컴포넌트로 병합
