+++ src/ui/MeetingDetail/CommentsList/index.tsx
@@ -1,18 +1,55 @@
 import { Avatar, Box, Flex } from '@chakra-ui/react';
+import { useState, useEffect } from 'react';
 
 import CommentEdit from '../CommentEdit';
 
-interface CommentsListDataProps {
+const DUMMY_COMMENTS_LIST_DATA = [
+  {
+    id: 1,
+    avatarURL: 'https://bit.ly/dan-abramov',
+    nickName: '김규란',
+    contents:
+      '백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세',
+  },
+  {
+    id: 2,
+    avatarURL: 'https://bit.ly/dan-abramov',
+    nickName: '김재현',
+    contents:
+      '백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세',
+  },
+  {
+    id: 3,
+    avatarURL: 'https://bit.ly/dan-abramov',
+    nickName: '백민종',
+    contents:
+      '백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세',
+  },
+  {
+    id: 4,
+    avatarURL: 'https://bit.ly/dan-abramov',
+    nickName: '동해물과',
+    contents:
+      '백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세 백두산이 마르고 닳도록 하느님이 보우하사 우리 나라 만세 무궁화 삼천리 화려 강산 대한 사람 대한으로 길이 보전하세',
+  },
+];
+
+interface commentsListDataProps {
   id: number;
   avatarURL: string;
   nickName: string;
   contents: string;
 }
-interface CommentsListPorps {
-  commentsListData: CommentsListDataProps[];
-}
 
-const CommentsList = ({ commentsListData }: CommentsListPorps) => {
+const CommentsList = () => {
+  const [commentsListData, setCommentsListDate] = useState<
+    commentsListDataProps[]
+  >([]);
+
+  useEffect(() => {
+    setCommentsListDate(DUMMY_COMMENTS_LIST_DATA);
+  }, []);
+
   return (
     <Box mt="1.5rem">
       <Box fontSize="lg" fontWeight={700}>
